move some syscalls to separate file and improve them
